Split up environment context
