--- PWManager.Application/Context/IApplicationEnvironment.cs
@@ -1,16 +1,0 @@
-﻿using PWManager.Domain.Entities;
-
-namespace PWManager.Application.Context; 
-
-public interface IApplicationEnvironment {
-
-    public bool IsDevelopmentMode { get; init; }
-    
-    public bool RunningSession { get; set; }
-
-    public User? CurrentUser { get; set; }
-    
-    public Group? CurrentGroup { get; set; }
-    
-    public string? EncryptionKey { get; set; }
-}
