+++ PWManager.Application/Context/ICliEnvironment.cs
@@ -1,0 +1,6 @@
+﻿namespace PWManager.Application.Context; 
+
+public interface ICliEnvironment {
+    public bool RunningSession { get; set; }
+    public string Prompt { get; }
+}
