+++ PWManager.Application/Context/ICryptEnvironment.cs
@@ -1,0 +1,5 @@
+﻿namespace PWManager.Application.Context; 
+
+public interface ICryptEnvironment {
+    public string? EncryptionKey { get; set; }
+}
