+++ PWManager.Application/Context/IDebugEnvironment.cs
@@ -1,0 +1,5 @@
+﻿namespace PWManager.Application.Context; 
+
+public interface IDebugEnvironment {
+    public bool IsDevelopmentMode { get; init; }
+}
