+++ PWManager.Application/Context/IUserEnvironment.cs
@@ -1,0 +1,9 @@
+﻿using PWManager.Domain.Entities;
+
+namespace PWManager.Application.Context; 
+
+public interface IUserEnvironment {
+    public User? CurrentUser { get; set; }
+    
+    public Group? CurrentGroup { get; set; }
+}
